+++ src/Query/Select.php
@@ -6,6 +6,7 @@ namespace Formal\AccessLayer\Query;
 use Formal\AccessLayer\{
     Query,
     Query\Parameter,
+    Query\Select\Direction,
     Row,
     Table\Name,
     Table\Column,
@@ -26,14 +27,35 @@ final class Select implements Query
     /** @var Sequence<Column\Name|Column\Name\Namespaced|Column\Name\Aliased> */
     private Sequence $columns;
     private Where $where;
+    /** @var ?array{Column\Name|Column\Name\Namespaced|Column\Name\Aliased, Direction} */
+    private ?array $orderBy;
+    /** @var ?positive-int */
+    private ?int $limit;
+    /** @var ?positive-int */
+    private ?int $offset;
 
-    private function __construct(Name|Name\Aliased $table, bool $lazy)
-    {
+    /**
+     * @param Sequence<Column\Name|Column\Name\Namespaced|Column\Name\Aliased> $columns
+     * @param ?array{Column\Name|Column\Name\Namespaced|Column\Name\Aliased, Direction} $orderBy
+     * @param ?positive-int $limit
+     * @param ?positive-int $offset
+     */
+    private function __construct(
+        Name|Name\Aliased $table,
+        bool $lazy,
+        Sequence $columns,
+        Where $where,
+        ?array $orderBy = null,
+        ?int $limit = null,
+        ?int $offset = null,
+    ) {
         $this->table = $table;
         $this->lazy = $lazy;
-        /** @var Sequence<Column\Name> */
-        $this->columns = Sequence::of();
-        $this->where = Where::everything();
+        $this->columns = $columns;
+        $this->where = $where;
+        $this->orderBy = $orderBy;
+        $this->limit = $limit;
+        $this->offset = $offset;
     }
 
     /**
@@ -41,7 +63,7 @@ final class Select implements Query
      */
     public static function from(Name|Name\Aliased $table): self
     {
-        return new self($table, false);
+        return new self($table, false, Sequence::of(), Where::everything());
     }
 
     /**
@@ -49,7 +71,7 @@ final class Select implements Query
      */
     public static function onDemand(Name|Name\Aliased $table): self
     {
-        return new self($table, true);
+        return new self($table, true, Sequence::of(), Where::everything());
     }
 
     /**
@@ -59,18 +81,60 @@ final class Select implements Query
         Column\Name|Column\Name\Namespaced|Column\Name\Aliased $first,
         Column\Name|Column\Name\Namespaced|Column\Name\Aliased ...$rest,
     ): self {
-        $self = clone $this;
-        $self->columns = Sequence::of($first, ...$rest);
-
-        return $self;
+        return new self(
+            $this->table,
+            $this->lazy,
+            Sequence::of($first, ...$rest),
+            $this->where,
+            $this->orderBy,
+            $this->limit,
+            $this->offset,
+        );
     }
 
     public function where(Specification $specification): self
     {
-        $self = clone $this;
-        $self->where = Where::of($specification);
+        return new self(
+            $this->table,
+            $this->lazy,
+            $this->columns,
+            Where::of($specification),
+            $this->orderBy,
+            $this->limit,
+            $this->offset,
+        );
+    }
 
-        return $self;
+    public function orderBy(
+        Column\Name|Column\Name\Namespaced|Column\Name\Aliased $column,
+        Direction $direction,
+    ): self {
+        return new self(
+            $this->table,
+            $this->lazy,
+            $this->columns,
+            $this->where,
+            [$column, $direction],
+            $this->limit,
+            $this->offset,
+        );
+    }
+
+    /**
+     * @param positive-int $limit
+     * @param positive-int $offset
+     */
+    public function limit(int $limit, int $offset = null): self
+    {
+        return new self(
+            $this->table,
+            $this->lazy,
+            $this->columns,
+            $this->where,
+            $this->orderBy,
+            $limit,
+            $offset,
+        );
     }
 
     public function parameters(): Sequence
@@ -82,10 +146,26 @@ final class Select implements Query
     {
         /** @var non-empty-string */
         return \sprintf(
-            'SELECT %s FROM %s %s',
+            'SELECT %s FROM %s %s%s%s%s',
             $this->columns->empty() ? '*' : $this->buildColumns(),
             $this->table->sql(),
             $this->where->sql(),
+            match ($this->orderBy) {
+                null => '',
+                default => \sprintf(
+                    ' ORDER BY %s %s',
+                    $this->orderBy[0]->sql(),
+                    $this->orderBy[1]->sql(),
+                ),
+            },
+            match ($this->limit) {
+                null => '',
+                default => ' LIMIT '.$this->limit,
+            },
+            match ($this->offset) {
+                null => '',
+                default => ' OFFSET '.$this->offset,
+            },
         );
     }